requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
azure-cognitiveservices-speech==1.26.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
distlib==0.3.6
filelock==3.10.7
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.2
platformdirs==3.2.0
PyAudio==0.2.13
pycparser==2.21
requests==2.28.2
sounddevice==0.4.6
tqdm==4.65.0
urllib3==1.26.15
virtualenv==20.21.0
yarl==1.8.2